Added OCR as tab on manifest
